% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_scl.R
\name{theme_scl}
\alias{theme_scl}
\title{SCL Theming Function}
\usage{
theme_scl(
  base_size = 12,
  base_family = "",
  plot_title_size = 20,
  plot_title_color = scl::scl_colors["logo"]
)
}
\arguments{
\item{base_size}{Size of font}

\item{base_family}{Sans or Serif}
}
\description{
SCL Theming Function
}
